some minor fixes, and one important 2/4 one in the sbreak function

+++ util/led/finish.c
@@ -188,7 +188,7 @@ put_locals(name, nnames)
 		}
 		iname++;
 	}
-	wr_name(name, oname - name);
+	wr_name(name, (unsigned int) (oname - name));
 }
 
 /*
